graph

+++ lib/roa_api/src/interface/camera.c
@@ -4,7 +4,7 @@
 
 void
 rep_camera_set(
-  uint64_t * object_ids,
+  uint32_t * object_ids,
   struct rep_camera_desc * desc,
   unsigned count)
 {
@@ -16,7 +16,7 @@ rep_camera_set(
 
 void
 rep_camera_get(
-  const uint64_t * object_ids,
+  const uint32_t * object_ids,
   struct rep_camera_desc * desc,
   unsigned count)
 {
